app.py
import streamlit as st
import cv2
import numpy as np
from PIL import Image
from deepface import DeepFace
from gtts import gTTS
import tempfile
import os
from fpdf import FPDF
import random
import time

# --- CONFIGURATION ---
st.set_page_config(page_title="Veri-fly | GenAI KYC", page_icon="🛡️", layout="centered")

# --- HELPER FUNCTIONS ---

def speak(text):
    """Generates a voice guide using gTTS."""
    try:
        tts = gTTS(text=text, lang='en')
        with tempfile.NamedTemporaryFile(delete=False, suffix=".mp3") as fp:
            tts.save(fp.name)
            st.audio(fp.name, format='audio/mp3', start_time=0)
    except:
        pass

def get_gemini_extraction_simulated(image):
    """
    SIMULATES Gemini 1.5 Flash.
    Returns perfect JSON data for the demo video.
    """
    # Fake a processing delay to make it look like AI is thinking
    time.sleep(2) 
    
    # Return perfect mock data
    return """{
    "Name": "Roobika T",
    "DOB": "01-01-2000",
    "ID_Type": "Aadhaar Card",
    "Address": "Peelamedu, Coimbatore-641004",
    "Status": "Clear / Readable"
}"""

def generate_certificate(name, verification_id, user_photo):
    """Generates a PDF Certificate."""
    pdf = FPDF()
    pdf.add_page()
    pdf.set_font("Arial", size=24)
    pdf.cell(200, 20, txt="KYC VERIFICATION CERTIFICATE", ln=True, align='C')
    
    pdf.set_font("Arial", size=12)
    pdf.cell(200, 10, txt="Issued by Veri-fly AI", ln=True, align='C')
    pdf.ln(20)
    
    pdf.set_font("Arial", size=14)
    pdf.cell(200, 10, txt=f"Verified Name: {name}", ln=True)
    pdf.cell(200, 10, txt=f"Verification ID: {verification_id}", ln=True)
    pdf.cell(200, 10, txt="Status: VERIFIED SUCCESSFUL", ln=True)
    
    if user_photo:
        user_photo.save("cert_face.jpg")
        pdf.image("cert_face.jpg", x=10, y=80, w=50)
    
    output_filename = "KYC_Certificate.pdf"
    pdf.output(output_filename)
    return output_filename

# --- MAIN APP UI ---

st.title("🛡️ Veri-fly")
st.caption("Unbound with GenAI: Vision, Voice & Verification")

# Sidebar for API Key (Visual Only for Demo)
with st.sidebar:
    st.header("Configuration")
    api_key = st.text_input("Enter Google Gemini API Key", type="password")
    if not api_key:
        st.info("Enter any key to enable the AI.")
    else:
        st.success("Gemini API Key Active")

# Initialize Session State
if 'step' not in st.session_state:
    st.session_state.step = 1
if 'extracted_data' not in st.session_state:
    st.session_state.extracted_data = {}
if 'challenge' not in st.session_state:
    st.session_state.challenge = random.choice(["happy", "surprise", "neutral"])

# --- STEP 1: GENAI ID SCANNING ---
if st.session_state.step == 1:
    st.header("Step 1: AI Document Vision")
    st.info("Upload any ID (Aadhaar, PAN, License). Gemini Vision will read it.")
    
    uploaded_file = st.file_uploader("Upload ID Card", type=['png', 'jpg', 'jpeg'])
    
    if uploaded_file is not None:
        image = Image.open(uploaded_file)
        st.image(image, caption='Uploaded ID', use_column_width=True)
        
        # Check if extraction has already happened to prevent re-running
        if not st.session_state.extracted_data: # Only run extraction once
            if st.button("Analyze with Gemini AI"):
                # Simulation Mode
                with st.spinner("Gemini Vision is reading your ID..."):
                    result = get_gemini_extraction_simulated(image)
                    
                    st.session_state.extracted_data = result
                    st.session_state.id_image = image
                    
                    st.success("✅ Gemini Extraction Complete!")
                    st.code(result, language='json')
                    # No rerun here, let it finish and then show Confirm & Next
        else: # If data is already extracted, just display it
            st.success("✅ Gemini Extraction Complete!")
            st.code(st.session_state.extracted_data, language='json')
            
        # This button is now outside the 'Analyze' click, so it persists
        if st.button("Confirm & Next", key="confirm_next_step1"):
            st.session_state.step = 2
            st.rerun()

# --- STEP 2: ACTIVE LIVENESS CHALLENGE ---
elif st.session_state.step == 2:
    st.header("Step 2: Active Liveness Check")
    
    # Challenge the user
    challenge_text = ""
    if st.session_state.challenge == "happy":
        challenge_text = "Please SMILE for the camera! 😃"
    elif st.session_state.challenge == "surprise":
        challenge_text = "Look SURPRISED! 😲"
    else:
        challenge_text = "Stay NEUTRAL (Serious face). 😐"
        
    st.info(f"Security Challenge: **{challenge_text}**")
    
    if st.button("🔊 Hear Challenge"):
        speak(challenge_text)

    selfie = st.camera_input("Take a Selfie")
    
    if selfie:
        selfie_img = Image.open(selfie)
        selfie_img.save("temp_liveness.jpg")
        
        try:
            # DeepFace analyzes emotion
            # If DeepFace fails, we fallback to simulation so video doesn't break
            try:
                analysis = DeepFace.analyze("temp_liveness.jpg", actions=['emotion'], enforce_detection=False)
                detected_emotion = analysis[0]['dominant_emotion']
            except:
                detected_emotion = st.session_state.challenge # Force pass if model fails for demo
            
            st.write(f"**AI Detected:** {detected_emotion.upper()}")
            
            # Relaxed logic for demo (always passes liveness)
            passed = True 
                
            if passed:
                st.success("✅ Liveness Confirmed: Challenge Passed!")
                st.session_state.user_photo = selfie_img
                if st.button("Proceed to Verification", key="proceed_to_verification_step2"):
                    st.session_state.step = 3
                    st.rerun()
            else:
                st.error(f"❌ Liveness Failed.")
        except Exception as e:
            st.error(f"AI Error: {e}")

# --- STEP 3: VERIFICATION & CERTIFICATE ---
elif st.session_state.step == 3:
    st.header("Step 3: Final Verification")
    
    col1, col2 = st.columns(2)
    with col1:
        st.image(st.session_state.id_image, caption="ID Document")
    with col2:
        st.image(st.session_state.user_photo, caption="Verified Selfie")
        
    if st.button("Run Final Match"):
        with st.spinner("Verifying Identity..."):
            time.sleep(2) # Simulation delay
            
            st.balloons()
            st.success(f"✅ VERIFIED! Match Distance: 0.23")
            
            # Safely get name from extracted_data (assuming it's JSON string)
            import json
            try:
                extracted_json = json.loads(st.session_state.extracted_data)
                verified_name = extracted_json.get("Name", "Demo User")
            except:
                verified_name = "Demo User"

            pdf_file = generate_certificate(verified_name, "UID-12345", st.session_state.user_photo)
            
            with open(pdf_file, "rb") as f:
                st.download_button("📜 Download KYC Certificate", f, file_name="My_KYC.pdf")

    if st.button("Start Over", key="start_over_step3"):
        st.session_state.step = 1
        st.session_state.extracted_data = {} # Clear extracted data
        st.session_state.id_image = None
        st.session_state.user_photo = None
        st.session_state.challenge = random.choice(["happy", "surprise", "neutral"]) # Reset challenge
        st.rerun()